replace constants from concrete cover check to the relevant eurocode directories + fix tests
